Rename package guards and split _compute/_discretization tests
Bring the first tranche of test modules into line with the AGENTS.md rule
that every source module `foo.py` keeps its tests in a sibling `foo_test.py`.

Package-scope guards move to `__init___test.py`, the one name that does
correspond to a real module for a test that covers a package as a whole:

  _compute/_layering_test.py   -> _compute/__init___test.py
  ion/_docstring_test.py       -> ion/__init___test.py
  channel/_docstring_test.py   -> channel/__init___test.py, with
  channel/_deprecation_test.py merged in (one package, one such file)

Splits by target module:

  _compute/topology_test.py    -> _compute/scheduling_test.py and
                                  _discretization/node_build_test.py
  _compute/spatial_params_test.py -> merged into _compute/layouts_test.py,
                                  where the asserted errors are raised
  _discretization/build_test.py -> _discretization/mechanism_test.py (12
                                  classes), geometry_test.py (8 classes),
                                  and base_test.py (LowerSmokeTest,
                                  LowerPropertyTest)

Shared fixtures move to the new _discretization/_testing.py, following the
vis/_testing.py precedent. The dead node_build import block in build_test.py
is dropped: all four names were imported and never used.

Straight renames with no content change:

  filter/filter_locset_test.py -> filter/locset_test.py
  io/asc/asc_test.py           -> io/asc/reader_test.py
  io/swc/swc_test.py           -> io/swc/reader_test.py

Collected test count is unchanged at 2276.
